+++ src/kubeseal_auto/cli.py
@@ -12,8 +12,8 @@ import click
 from icecream import ic
 
 from kubeseal_auto import __version__, console
+from kubeseal_auto.core.kubeseal import Kubeseal
 from kubeseal_auto.exceptions import ClusterConnectionError, SecretParsingError
-from kubeseal_auto.kubeseal import Kubeseal
 from kubeseal_auto.models import SecretParams, SecretType
