+++ src/kubeseal_auto/core/__init__.py
@@ -1,0 +1,15 @@
+"""Core infrastructure subpackage.
+
+This package contains the main Kubeseal facade class along with
+cluster and host management utilities.
+"""
+
+from kubeseal_auto.core.cluster import Cluster
+from kubeseal_auto.core.host import Host
+from kubeseal_auto.core.kubeseal import Kubeseal
+
+__all__ = [
+    "Cluster",
+    "Host",
+    "Kubeseal",
+]
